Fix KeyError in channel migration due to missing check

+++ plugins/layout.py
@@ -967,27 +967,29 @@ class Migrate(commands.GroupCog, name="migrate"):
         if channels:
             for _category, _channels in guild.by_category():
                 if _category:
-                    category_id = channels[DB_CHANNEL_NAME_MAP[_category.name]]
-                    for _channel in _category.channels:
-                        channel_id = channels[DB_CHANNEL_NAME_MAP[_channel.name]]
-                        if _channel.id != channel_id:
-                            if _channel not in [
-                                guild.rules_channel,
-                                guild.public_updates_channel,
-                            ]:
-                                await _channel.delete()
+                    if DB_CHANNEL_NAME_MAP[_category.name] in channels:
+                        category_id = channels[DB_CHANNEL_NAME_MAP[_category.name]]
+                        for _channel in _category.channels:
+                            channel_id = channels[DB_CHANNEL_NAME_MAP[_channel.name]]
+                            if _channel.id != channel_id:
+                                if _channel not in [
+                                    guild.rules_channel,
+                                    guild.public_updates_channel,
+                                ]:
+                                    await _channel.delete()
 
-                    if _category.id != category_id:
-                        await _category.delete()
+                        if _category.id != category_id:
+                            await _category.delete()
                 else:
                     for _channel in _channels:
-                        channel_id = channels[DB_CHANNEL_NAME_MAP[_channel.name]]
-                        if _channel.id != channel_id:
-                            if _channel not in [
-                                guild.rules_channel,
-                                guild.public_updates_channel,
-                            ]:
-                                await _channel.delete()
+                        if DB_CHANNEL_NAME_MAP[_channel.name] in channels:
+                            channel_id = channels[DB_CHANNEL_NAME_MAP[_channel.name]]
+                            if _channel.id != channel_id:
+                                if _channel not in [
+                                    guild.rules_channel,
+                                    guild.public_updates_channel,
+                                ]:
+                                    await _channel.delete()
 
         for _channel in guild.channels:
             if _channel.name not in DB_CHANNEL_NAME_MAP.keys():
